refactor(api): extract request logging helpers
